Move secondary Scientific related tasks to new package
Changes:
- Move remaing Task concerning ScientificStudy to the new package
- Convert Tasks to ExperienceImpact
- Move associated MetaTask classes
- InviteStudyCollaborator sends invites one at a time instead of all at the same time
- Add a new Skill Organisation that represents how organised a person is
- InviteStudyCollaborator, RespondToStudyInvite uses the new Organisation skill
- Convert Scientific Tasks to ExperienceImpact
- Create UnitTests for each Scientific Task
- UnitTest use a Chain of Responsibility pattern to create the Study in the correct phase
- MarsSimContext is a new interface for UnitTests to provide context to create simualtion entities
- Mars sim unit test has method to execute a Task for a maximum duration of time
- LabTask provides a centralised determineStudy method
- Remove unused methods from ChainOfCommand class
- Change method in BuildingManager to use ScienceType instead of ScientificStudy
- AssistScientificStudyResearch task does not use computing resource as it provides assistance to main Task

Part of #1285 & #1286
