+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/task/util/MetaTaskUtil.java
@@ -17,9 +17,7 @@ import java.util.stream.Collectors;
 import com.mars_sim.core.Simulation;
 import com.mars_sim.core.activities.GroupActivityMetaTask;
 import com.mars_sim.core.person.ai.task.meta.AnalyzeMapDataMeta;
-import com.mars_sim.core.person.ai.task.meta.AssistScientificStudyResearcherMeta;
 import com.mars_sim.core.person.ai.task.meta.BudgetResourcesMeta;
-import com.mars_sim.core.person.ai.task.meta.CompileScientificStudyResultsMeta;
 import com.mars_sim.core.person.ai.task.meta.ConnectOnlineMeta;
 import com.mars_sim.core.person.ai.task.meta.ConsolidateContainersMeta;
 import com.mars_sim.core.person.ai.task.meta.ConstructBuildingMeta;
@@ -31,7 +29,6 @@ import com.mars_sim.core.person.ai.task.meta.DigLocalRegolithMeta;
 import com.mars_sim.core.person.ai.task.meta.DoInventoryMeta;
 import com.mars_sim.core.person.ai.task.meta.EatDrinkMeta;
 import com.mars_sim.core.person.ai.task.meta.ExamineBodyMeta;
-import com.mars_sim.core.person.ai.task.meta.InviteStudyCollaboratorMeta;
 import com.mars_sim.core.person.ai.task.meta.ListenToMusicMeta;
 import com.mars_sim.core.person.ai.task.meta.LoadVehicleMeta;
 import com.mars_sim.core.person.ai.task.meta.MaintainBuildingMeta;
@@ -41,23 +38,17 @@ import com.mars_sim.core.person.ai.task.meta.ManufactureGoodMeta;
 import com.mars_sim.core.person.ai.task.meta.MeetTogetherMeta;
 import com.mars_sim.core.person.ai.task.meta.ObserveAstronomicalObjectsMeta;
 import com.mars_sim.core.person.ai.task.meta.OptimizeSystemMeta;
-import com.mars_sim.core.person.ai.task.meta.PeerReviewStudyPaperMeta;
-import com.mars_sim.core.science.task.PerformLaboratoryResearchMeta;
-import com.mars_sim.core.science.task.PerformMathematicalModelingMeta;
-import com.mars_sim.core.science.task.StudyFieldSamplesMeta;
 import com.mars_sim.core.person.ai.task.meta.PlanMissionMeta;
 import com.mars_sim.core.person.ai.task.meta.PlayHoloGameMeta;
 import com.mars_sim.core.person.ai.task.meta.PrepareDessertMeta;
 import com.mars_sim.core.person.ai.task.meta.PrescribeMedicationMeta;
 import com.mars_sim.core.person.ai.task.meta.ProduceFoodMeta;
-import com.mars_sim.core.person.ai.task.meta.ProposeScientificStudyMeta;
 import com.mars_sim.core.person.ai.task.meta.ReadMeta;
 import com.mars_sim.core.person.ai.task.meta.RecordActivityMeta;
 import com.mars_sim.core.person.ai.task.meta.RelaxMeta;
 import com.mars_sim.core.person.ai.task.meta.RepairMalfunctionMeta;
 import com.mars_sim.core.person.ai.task.meta.ReportMissionControlMeta;
 import com.mars_sim.core.person.ai.task.meta.RequestMedicalTreatmentMeta;
-import com.mars_sim.core.person.ai.task.meta.RespondToStudyInvitationMeta;
 import com.mars_sim.core.person.ai.task.meta.RestingMedicalRecoveryMeta;
 import com.mars_sim.core.person.ai.task.meta.ReturnLightUtilityVehicleMeta;
 import com.mars_sim.core.person.ai.task.meta.ReviewJobReassignmentMeta;
@@ -81,6 +72,15 @@ import com.mars_sim.core.person.ai.task.util.MetaTask.TaskScope;
 import com.mars_sim.core.person.ai.task.util.MetaTask.WorkerType;
 import com.mars_sim.core.robot.ai.task.ChargeMeta;
 import com.mars_sim.core.robot.ai.task.SavePowerMeta;
+import com.mars_sim.core.science.task.AssistScientificStudyResearcherMeta;
+import com.mars_sim.core.science.task.CompileScientificStudyResultsMeta;
+import com.mars_sim.core.science.task.InviteStudyCollaboratorMeta;
+import com.mars_sim.core.science.task.PeerReviewStudyPaperMeta;
+import com.mars_sim.core.science.task.PerformLaboratoryResearchMeta;
+import com.mars_sim.core.science.task.PerformMathematicalModelingMeta;
+import com.mars_sim.core.science.task.ProposeScientificStudyMeta;
+import com.mars_sim.core.science.task.RespondToStudyInvitationMeta;
+import com.mars_sim.core.science.task.StudyFieldSamplesMeta;
 
 /**
  * A utility task for getting the list of meta tasks.
