+++ mars-sim-core/src/main/java/com/mars_sim/core/science/task/PerformLaboratoryResearch.java
@@ -6,8 +6,7 @@
  */
 package com.mars_sim.core.science.task;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.HashSet;
 import java.util.Set;
 
 import com.mars_sim.core.person.Person;
@@ -16,7 +15,6 @@ import com.mars_sim.core.person.ai.task.util.ExperienceImpact;
 import com.mars_sim.core.person.ai.task.util.TaskPhase;
 import com.mars_sim.core.science.ScienceType;
 import com.mars_sim.core.science.ScientificStudy;
-import com.mars_sim.core.structure.Lab;
 import com.mars_sim.tools.Msg;
 import com.mars_sim.tools.util.RandomUtil;
 
@@ -34,6 +32,14 @@ public class PerformLaboratoryResearch extends LabTask {
 	/** Task phases. */
 	private static final TaskPhase RESEARCHING = new TaskPhase(Msg.getString("Task.phase.researching")); //$NON-NLS-1$
 
+	private static final Set<ScienceType> LAB_SCIENCE = new HashSet<>();
+
+	static {
+		// All science minus the modelling sciences
+		LAB_SCIENCE.addAll(Set.of(ScienceType.values()));
+		LAB_SCIENCE.removeAll(PerformMathematicalModeling.MODELLING_SCIENCE);
+	}
+
 
 	/**
 	 * Create a Task to perform lab research. This will select the most appropirate Scientific Study for the Person
@@ -42,7 +48,7 @@ public class PerformLaboratoryResearch extends LabTask {
 	 * @return
 	 */
 	public static PerformLaboratoryResearch createTask(Person person) {
-		var study = determineStudy(person);
+		var study = determineStudy(person, LAB_SCIENCE);
 		if ((study != null) && (study.getContribution(person) != null)) {
 			// Found a suitable study
 			var impact = new ExperienceImpact(25D, NaturalAttributeType.ACADEMIC_APTITUDE,
@@ -64,48 +70,5 @@ public class PerformLaboratoryResearch extends LabTask {
 		super(NAME, person, study, impact, 10D + RandomUtil.getRandomDouble(50D),
 				RESEARCHING, "Task.description.performLaboratoryResearch.detail");
 
-	}
-
-	/**
-	 * Determines the scientific study that will be researched.
-	 * 
-	 * @return study or null if none available.
-	 */
-	private static ScientificStudy determineStudy(Person person) {
-		List<ScientificStudy> possibleStudies = new ArrayList<>();
-
-		// Add primary study if in research phase.
-		ScientificStudy primaryStudy = person.getStudy();
-		if ((primaryStudy != null) && ScientificStudy.RESEARCH_PHASE.equals(primaryStudy.getPhase())
-					&& !primaryStudy.isPrimaryResearchCompleted()) {
-
-			// Check that a lab is available for primary study science.
-			Lab lab = getLocalLab(person, primaryStudy.getScience());
-			if (lab != null) {
-
-				// Primary study added twice to double chance of random selection.
-				possibleStudies.add(primaryStudy);
-				possibleStudies.add(primaryStudy);
-			}
-		}
-
-		// Add all collaborative studies in research phase.
-		for(ScientificStudy collabStudy : person.getCollabStudies()) {
-			if (ScientificStudy.RESEARCH_PHASE.equals(collabStudy.getPhase())
-					&& !collabStudy.isCollaborativeResearchCompleted(person)) {
-
-				// Check that a lab is available for collaborative study science.
-				ScienceType collabScience = collabStudy.getContribution(person);
-
-				Lab lab = getLocalLab(person, collabScience);
-				if (lab != null) {
-
-					possibleStudies.add(collabStudy);
-				}
-			}
-		}
-
-		// Randomly select study.
-		return RandomUtil.getRandomElement(possibleStudies);
 	}
 }
