+++ mars-sim-core/src/main/resources/messages.properties
@@ -1342,6 +1342,7 @@ SkillType.piloting         = Piloting
 SkillType.preparingDessert = Preparing Dessert
 SkillType.prospecting      = Prospecting
 SkillType.psychology       = Psychology
+SkillType.organisation     = Organisation
 SkillType.reporting 	   = Reporting
 SkillType.trading          = Trading
